refactor `test_tasks.py` into multiple test modules
