+++ tests/message/tasks/test_waiting_for_results.py
@@ -1,0 +1,15 @@
+import unittest
+
+from golem_messages import factories
+from golem_messages import message
+from tests.message import mixins
+
+
+class WaitingForResultsTest(
+        mixins.RegisteredMessageTestMixin,
+        mixins.SerializationMixin,
+        mixins.TaskIdMixin,
+        unittest.TestCase):
+    FACTORY = factories.tasks.WaitingForResultsFactory
+    MSG_CLASS = message.tasks.WaitingForResults
+    TASK_ID_PROVIDER = 'task_to_compute'
